Add files via upload
The functional task of the EFSH module is the transformation of the wave functions of atomic ligands in the transition from atomic to the unified molecular coordinate system. This transition presupposes the wave function expansion of the electrons of the atomic ligand in a series in terms of spherical harmonics relative to the selected center of the molecular system. The generated EFSH functions describe higher harmonics of molecular orbitals that are localized in the regions with atomic ligands.