XMHFL
The main functional task of the XMHFL software is to calculate the wave functions of the states of electrons of a molecular system, describing different degrees of excitation by X-rays. XMHFL provides the ability to account for emerging vacancies in the deep, valent, and subvalent electron shells of a molecular system. Taking into account the emerging vacancy provides a correct description of the relaxation of the electron density of the molecular system. This significantly affects the accuracy of calculating the observed spectroscopic quantities of the molecular system.